Allow a shared folder to be disabled
Allow a shared folder to be disabled by a flag called "disabled".
This is similar to what the Virtualbox provider supports [1] and to
what the VMware provider [2] already supports.

This feature is especially useful for cases where you do not want to
rsync tons of data to all machines in a multi-VM config.

This will be unnecessary once support for this is included in
Vagrant core [2].

[1]: https://github.com/mitchellh/vagrant/commit/d18d1d5a71bc17e58127b56c521143dd4965f628
[2]: https://github.com/mitchellh/vagrant/issues/1004#issuecomment-15032502

Acked-by: Marc-André Courtois <macourtois@iweb.com>