Handle not having the box and providing it via a box_url.
The way this was implemented breaks with a terrible runtime error if you don't have the box
already installed nor a box_url in your vagrant file. However, this permits you to provide
the openstack api_key and auth_url directly in your box file. This wouldn't be possible if
you validated the config prior to handling the box url. The proper way to fix the terrible
runtime error would be to check for box_url directly in the HandleBoxUrl method in Vagrant
core.

Reviewed-by: Philippe Godin <godp1301@gmail.com>